# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraFleet API",
        "description": "Fleet management system API",
        "contact": {
            "email": "admin@larafleet.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Lista veicoli",
                "operationId": "f627fb85543e356532922246d50e6e2a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "in_use",
                                "maintenance"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "car",
                                "truck",
                                "scooter",
                                "van",
                                "bus"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginata di veicoli"
                    },
                    "401": {
                        "description": "Non autenticato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Crea veicolo (admin)",
                "operationId": "3b250a933fc5648aedd688fdcf4eae81",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plate",
                                    "brand",
                                    "model",
                                    "year",
                                    "type"
                                ],
                                "properties": {
                                    "plate": {
                                        "type": "string",
                                        "example": "AB123CD"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Fiat"
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Panda"
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2022
                                    },
                                    "km": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "car",
                                            "truck",
                                            "scooter",
                                            "van",
                                            "bus"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "in_use",
                                            "maintenance"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Veicolo creato"
                    },
                    "403": {
                        "description": "Non autorizzato"
                    },
                    "422": {
                        "description": "Errore di validazione"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Dettaglio veicolo",
                "operationId": "00b76a563ae78ef08e92480a3cd58a9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dettaglio veicolo"
                    },
                    "404": {
                        "description": "Veicolo non trovato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Aggiorna veicolo (admin)",
                "operationId": "322797dc21f323e627fc60773b7130c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plate": {
                                        "type": "string"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "km": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "car",
                                            "truck",
                                            "scooter",
                                            "van",
                                            "bus"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "available",
                                            "in_use",
                                            "maintenance"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Veicolo aggiornato"
                    },
                    "403": {
                        "description": "Non autorizzato"
                    },
                    "404": {
                        "description": "Veicolo non trovato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Elimina veicolo (admin)",
                "operationId": "7fbf332babe156123bfa2b84d32bb68a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Veicolo eliminato"
                    },
                    "403": {
                        "description": "Non autorizzato"
                    },
                    "404": {
                        "description": "Veicolo non trovato"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}